Integrate account system, fix linking errors

+++ include/utils.h
@@ -8,7 +8,7 @@
 using namespace std;
 
 int _random();
-vector<string> fitStringToWidth(string str, int width);
+vector<string> fitStringToWidth(string str, size_t width);
 
 #define RANDOM(min, max) (_random()%(int)(max-min)+min) // random with min and max boundaries
 #define RANDOM_POSITIVE(min, max) ((_random()>>1)%(max-min)+min) // remove the signedness bit
